+++ src/main/java/org/folio/service/processing/ParallelFileChunkingProcessor.java
@@ -1,5 +1,16 @@
 package org.folio.service.processing;
 
+import static org.folio.rest.RestVerticle.MODULE_SPECIFIC_ARGS;
+import static org.folio.rest.jaxrs.model.StatusDto.ErrorStatus.FILE_PROCESSING_ERROR;
+import static org.folio.rest.jaxrs.model.StatusDto.Status.ERROR;
+import static org.folio.rest.jaxrs.model.UploadDefinition.Status.COMPLETED;
+
+import java.io.File;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
+
 import io.vertx.core.CompositeFuture;
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
@@ -27,17 +38,6 @@ import org.folio.service.storage.FileStorageService;
 import org.folio.service.storage.FileStorageServiceBuilder;
 import org.folio.service.upload.UploadDefinitionService;
 import org.folio.service.upload.UploadDefinitionServiceImpl;
-
-import java.io.File;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.concurrent.atomic.AtomicBoolean;
-
-import static org.folio.rest.RestVerticle.MODULE_SPECIFIC_ARGS;
-import static org.folio.rest.jaxrs.model.StatusDto.ErrorStatus.FILE_PROCESSING_ERROR;
-import static org.folio.rest.jaxrs.model.StatusDto.Status.ERROR;
-import static org.folio.rest.jaxrs.model.UploadDefinition.Status.COMPLETED;
 
 /**
  * Processing files in parallel threads, one thread per one file.
@@ -138,7 +138,11 @@ public class ParallelFileChunkingProcessor implements FileProcessor {
           coordinator.acceptLock();
           List<String> records = reader.next();
           recordsCounter.add(records.size());
-          RawRecordsDto chunk = new RawRecordsDto().withRecords(records).withCounter(recordsCounter.getValue()).withLast(false);
+          RawRecordsDto chunk = new RawRecordsDto()
+            .withRecords(records)
+            .withContentType(reader.getContentType())
+            .withCounter(recordsCounter.getValue())
+            .withLast(false);
           chunkSentFutures.add(postRawRecords(fileDefinition.getJobExecutionId(), chunk, canSendNextChunk, coordinator, params));
         } else {
           String errorMessage = "Can not send next chunk of file " + fileDefinition.getSourcePath();
@@ -153,7 +157,10 @@ public class ParallelFileChunkingProcessor implements FileProcessor {
           resultFuture.fail(errorMessage);
         } else {
           // Sending the last chunk
-          RawRecordsDto chunk = new RawRecordsDto().withCounter(recordsCounter.getValue()).withLast(true);
+          RawRecordsDto chunk = new RawRecordsDto()
+            .withContentType(reader.getContentType())
+            .withCounter(recordsCounter.getValue())
+            .withLast(true);
           postRawRecords(fileDefinition.getJobExecutionId(), chunk, canSendNextChunk, coordinator, params).setHandler(r -> {
             if (r.failed()) {
               String errorMessage = "File processing stopped. Can not send the last chunk of the file " + fileDefinition.getSourcePath();